+++ src/memory/memset.c
@@ -8,7 +8,7 @@ EXPORT(memset);
 
 // https://en.cppreference.com/w/c/string/byte/memset
 void *memset(void *dest, int ch, size_t n) {
-    unsigned char *ptr = (unsigned char*)dest;
+    unsigned char *ptr = dest;
     size_t i;
 
     for (i = 0; i < n; i++, ptr++) {