+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `libc` functions:
+  - Memory functions: `memcpy`.
+
 ## [v0.1.0] - 2023-11-07
 
 ### Added